Move iter_except to utils as it is a generally useful function
